Move common models to core/models.py
