+++ codegen/specs/stytch/sessions.yml
@@ -3,7 +3,7 @@ additional_imports:
   - import jwt
   - import datetime
   - import time
-  - from stytch.models.common import StytchSession
+  - from stytch.core.models import StytchSession
 methods:
   - name: get
     use_base_path_as_api_path: True
