+++ codegen/types/templates/api_responses.tmpl
@@ -3,7 +3,7 @@
 from typing import Any, Dict, List, Optional
 
 
-from stytch.models.common import (
+from stytch.core.models import (
   AuthenticationFactor,
   BiometricRegistration,
   CryptoWallet,
