+++ stytch/models/passwords.py
@@ -2,7 +2,7 @@
 
 from typing import Any, Dict, Optional
 
-from stytch.models.common import ResponseBase, StytchSession, User
+from stytch.core.models import ResponseBase, StytchSession, User
 
 
 class CreateResponse(ResponseBase):
