+++ stytch/models/passwords_session.py
@@ -2,7 +2,7 @@
 
 from typing import Optional
 
-from stytch.models.common import ResponseBase, StytchSession
+from stytch.core.models import ResponseBase, StytchSession
 
 
 class ResetResponse(ResponseBase):
